A_S_bot/question_scraper_newDB/requirements.txt
requests==2.31.0
Pillow==11.0.0
pytesseract==0.3.10
pynput==1.7.6
opencv-python==4.12.0.88
numpy==2.2.6
